requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
openai==1.54.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0